+++ main.c
@@ -8,6 +8,10 @@
 #include "quickresponse.h"
 #include "shared.h"
 
+/* NOTE: not to be confused with MAX_VERSION */
+#define VERSION_LIMIT 5
+
+/* Terminal argument */
 typedef enum targ_e
 {
   ARG_NONE     = 0,
@@ -18,6 +22,7 @@ typedef enum targ_e
   ARG_INFO     = 0x10,
   ARG_OPTIMIZE = 0x20,
   ARG_HELP     = 0x40,
+  ARG_NOLIMIT  = 0x80,
   ARG_RESERVED __attribute__((unavailable("bit mask limit"))) = 0x8000,
   ARG_MASK,
   ARG_VER,
@@ -26,18 +31,18 @@ typedef enum targ_e
   ARG_BMP,
   ARG_SVG
 } targ_t;
-#define NUM_ARGS      13
+#define NUM_ARGS      14
 #define NUM_MANDATORY 1
 
 static const char* args_[NUM_ARGS] = {
-  "--nocopy", "--verbose", "--raw",
-  "--noinline", "--version", "--optimize", "--help",
+  "--nocopy", "--verbose", "--raw", "--noinline",
+  "--version", "--optimize", "--help", "--nolimit",
   "-m", "-u", "-l", "-s", "-B", "-K"
 };
 
 static const targ_t arge_[NUM_ARGS] = {
-  ARG_NOCOPY, ARG_VERBOSE, ARG_RAW,
-  ARG_NOINLINE, ARG_INFO, ARG_OPTIMIZE, ARG_HELP,
+  ARG_NOCOPY, ARG_VERBOSE, ARG_RAW, ARG_NOINLINE,
+  ARG_INFO, ARG_OPTIMIZE, ARG_HELP, ARG_NOLIMIT,
   ARG_MASK, ARG_VER, ARG_LEVEL, ARG_SCALE,
   ARG_BMP, ARG_SVG
 };
@@ -59,7 +64,7 @@ phelp_(const char* __restrict__ cmdln)
     "  -m <uint>    force choice of mask <0-7>, regardless of penalty" _nl
     "  -s <uint>    scale image output <1-" _xstr(MAX_SCALE) "> times" _nl
     "  -u <uint>    force use of version <1-" _xstr(MAX_VERSION) "> code"
-                      "(or lower, if" _nl
+                      " (or lower, if" _nl
     "                 used with --optimize)" _nl
     "  -B <string>  create bitmap file with generated code" _nl
     "  -K <string>  create scalable vector image, disregards -s" _nl,
@@ -71,7 +76,7 @@ int
 main(const int argc, char* argv[])
 {
 #if defined(_WIN32)
-  // NOTE: to allow box-drawing characters
+  /* NOTE: to allow box-drawing characters */
   system("chcp 65001>nul");
 #endif
 
@@ -206,8 +211,17 @@ main(const int argc, char* argv[])
     }
     if (!(options & ARG_NOINLINE))
     {
-      pdebug("printing inline qrcode");
-      qrcode_print(qr, options & ARG_RAW);
+      const uint8_t version = qrcode_version(qr);
+      if (version > VERSION_LIMIT && !(options & ARG_NOLIMIT))
+      {
+        eprintf("could not print inline qrcode, version %d too high,"
+                "use --nolimit and try again", version);
+      }
+      else
+      {
+        pdebug("printing inline qrcode");
+        qrcode_print(qr, options & ARG_RAW);
+      }
     }
     if (imgout != NULL)
     {