+++ quickresponse.h
@@ -1,15 +1,20 @@
+#include <stdint.h>
 #ifndef QUICKRESPONSE_H
 #define QUICKRESPONSE_H 1
 
 #include "mask.h"
 #include "shared.h"
 
+/* Image format */
 typedef enum imgfmt_e
 {
   FMT_BMP,
   FMT_SVG
 } __attribute__((packed)) imgfmt_t;
 
+/* Creates QR code from string,
+   allows print/output of generated image
+*/
 typedef struct qrcode_s qrcode_t;
 
 __attribute__((__nonnull__)) int
@@ -20,6 +25,9 @@ delete_qrcode(qrcode_t** self);
 
 __attribute__((__nonnull__)) int
 qrcode_forcemask(qrcode_t* self, int mask);
+
+__attribute__((__nonnull__)) uint8_t
+qrcode_version(const qrcode_t* self);
 __attribute__((__nonnull__)) void
 qrcode_print(const qrcode_t* self, bool useraw);
 __attribute__((__nonnull__)) int